feat: refactor SkillDetailContent and introduce new components for improved structure and functionality
