+++ app/web/pages/skills/detail/style.scss
@@ -7,7 +7,7 @@
     background: #f7f9fb;
     font-family: var(--skill-detail-font-body);
     color: #2a3439;
-    .skill-detail-figma-icon {
+    .skill-detail-icon {
         display: inline-block;
         object-fit: contain;
         flex: 0 0 auto;
